rest-api/function-source-DIABETES-PREDICTOR/requirements.txt
numpy==1.26.2
pandas==2.1.4
Requests==2.31.0
scikit_learn==1.3.2
Flask==2.3.3
google-cloud-storage